Enables support for rouge and pygments.
To enable highlight.js the user must set the site parameter useHighlightjs
and optionally set adocHighlightTheme to one of the bundled themes,
default is github.
Otherwise, user is free to supply a custom_highlight.css file and that
will be loaded.

By not setting the highlight.js options, the user should then set the
[markup.asciidocext.attributes] to let asciidoctor handle syntax
highlighting by itself either with rouge or pygments.rb.
Under such condition, the user can set the attribute '*-css' to style
to let asciidoctor inline styles or set it to 'class' and provide a
custom_highlight.css file containing the definitions it needs.
The css file can be generated with rougify or pygmentize and they can be
made compatible between each other.